feat: usando contexto dentro do NewCycleForm
