+++ ignite-reactjs-2023/03-ignite-timer/src/pages/Home/components/Countdown/index.tsx
@@ -1,11 +1,10 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext, useEffect } from "react";
 import { CountdownContainer, Separator } from "./styles";
 import { differenceInSeconds } from "date-fns";
 import { CyclesContext } from "../..";
 
 export function Countdown(){
-    const { activeCycle, activeCycleId, markCurrentCycleAsFinished } = useContext(CyclesContext);
-    const [amountSecondsPassed, setAmountSecondsPassed] = useState(0);
+    const { activeCycle, activeCycleId, markCurrentCycleAsFinished, amountSecondsPassed, setSecondsPassed } = useContext(CyclesContext);
     const totalSeconds = activeCycle ? activeCycle.minutesAmount * 60 : 0;
 
     useEffect(() => {
@@ -25,10 +24,10 @@ export function Countdown(){
                     //         }
                     //     }),
                     // )
-                    setAmountSecondsPassed(totalSeconds);
+                    setSecondsPassed(totalSeconds);
                     clearInterval(interval);
                 }else{
-                    setAmountSecondsPassed(secondsDifference);
+                    setSecondsPassed(secondsDifference);
                 }
             }, 1000);
         }
@@ -37,7 +36,7 @@ export function Countdown(){
             clearInterval(interval);
         }
 
-    }, [activeCycle, totalSeconds, activeCycleId, markCurrentCycleAsFinished]);
+    }, [activeCycle, totalSeconds, activeCycleId, setSecondsPassed, markCurrentCycleAsFinished]);
 
     const currentSeconds = activeCycle ? totalSeconds - amountSecondsPassed : 0;
